+++ include/quantum_state/Psi.hpp
@@ -36,7 +36,7 @@ public:
     static constexpr unsigned int  max_M = MAX_HIDDEN_SPINS;
 
     unsigned int    num_active_params;
-    double          prefactor;
+    float          prefactor;
 
     complex_t* a;
     complex_t* b;
@@ -65,7 +65,7 @@ public:
 
     HDINLINE
     complex_t log_psi_s(const Spins& spins) const {
-        complex_t result(0.0, 0.0);
+        complex_t result(0.0f, 0.0f);
         for(unsigned int i = 0; i < this->N; i++) {
             result += this->a[i] * spins[i];
         }
@@ -86,15 +86,15 @@ public:
         #ifdef __CUDA_ARCH__
 
         auto summand = complex_t(
-            (threadIdx.x < this->N ? this->a[threadIdx.x] * spins[threadIdx.x] : complex_t(0.0, 0.0)) +
-            (threadIdx.x < this->M ? my_logcosh(angles[threadIdx.x]) : complex_t(0.0, 0.0))
+            (threadIdx.x < this->N ? this->a[threadIdx.x] * spins[threadIdx.x] : complex_t(0.0f, 0.0f)) +
+            (threadIdx.x < this->M ? my_logcosh(angles[threadIdx.x]) : complex_t(0.0f, 0.0f))
         );
 
         tree_sum(result, this->M, summand);
 
         #else
 
-        result = complex_t(0.0, 0.0);
+        result = complex_t(0.0f, 0.0f);
         for(auto i = 0u; i < this->N; i++) {
             result += this->a[i] * spins[i];
         }
@@ -106,22 +106,22 @@ public:
     }
 
     HDINLINE
-    void log_psi_s_real(double& result, const Spins& spins, const Angles& angles) const {
+    void log_psi_s_real(float& result, const Spins& spins, const Angles& angles) const {
         // CAUTION: 'result' has to be a shared variable.
         // j = threadIdx.x
 
         #ifdef __CUDA_ARCH__
 
-        auto summand = double(
-            (threadIdx.x < this->N ? this->a[threadIdx.x].real() * spins[threadIdx.x] : 0.0) +
-            (threadIdx.x < this->M ? my_logcosh(angles[threadIdx.x]).real() : 0.0)
+        auto summand = float(
+            (threadIdx.x < this->N ? this->a[threadIdx.x].real() * spins[threadIdx.x] : 0.0f) +
+            (threadIdx.x < this->M ? my_logcosh(angles[threadIdx.x]).real() : 0.0f)
         );
 
         tree_sum(result, this->M, summand);
 
         #else
 
-        result = 0.0;
+        result = 0.0f;
         for(auto i = 0u; i < this->N; i++) {
             result += this->a[i].real() * spins[i];
         }
@@ -136,7 +136,7 @@ public:
         const unsigned int j, const unsigned int position, const Spins& new_spins, Angles& angles
     ) const {
         if(j < this->get_num_angles()) {
-            angles[j] += 2.0 * new_spins[position] * this->W[position * this->M + j];
+            angles[j] += 2.0f * new_spins[position] * this->W[position * this->M + j];
         }
     }
 
@@ -152,22 +152,22 @@ public:
 
 #endif // __CUDACC__
 
-    // complex<double> psi_s_std(const Spins& spins) const {
+    // complex<float> psi_s_std(const Spins& spins) const {
     //     return this->psi_s(spins).to_std();
     // }
 
     // HDINLINE
-    // double probability_s(const Spins& spins, const Angles& angles) const {
-    //     return exp(2.0 * (log(this->prefactor) + this->log_psi_s(spins, angles).real()));
+    // float probability_s(const Spins& spins, const Angles& angles) const {
+    //     return exp(2.0f * (log(this->prefactor) + this->log_psi_s(spins, angles).real()));
     // }
 
-    // double probability_s_py(const Spins& spins) const {
+    // float probability_s_py(const Spins& spins) const {
     //     return this->probability_s(spins);
     // }
 
     HDINLINE
-    double probability_s(const double log_psi_s_real) const {
-        return exp(2.0 * (log(this->prefactor) + log_psi_s_real));
+    float probability_s(const float log_psi_s_real) const {
+        return exp(2.0f * (log(this->prefactor) + log_psi_s_real));
     }
 
     HDINLINE
@@ -224,7 +224,7 @@ public:
         const PsiDerivatives& psi_derivatives
     ) const {
         if(k < this->N) {
-            return complex_t(spins[k], 0.0);
+            return complex_t(spins[k], 0.0f);
         }
 
         const auto N_plus_M = this->N + this->M;
@@ -277,25 +277,25 @@ public:
 
 public:
     Psi(const unsigned int N, const unsigned int M, const bool gpu=true);
-    Psi(const unsigned int N, const unsigned int M, const int seed=0, const double noise=1e-4, const bool gpu=true);
+    Psi(const unsigned int N, const unsigned int M, const int seed=0, const float noise=1e-4, const bool gpu=true);
     Psi(
         const unsigned int N,
         const unsigned int M,
-        const std::complex<double>* a,
-        const std::complex<double>* b,
-        const std::complex<double>* W,
-        const std::complex<double>* n,
-        const double prefactor=1.0,
+        const std::complex<float>* a,
+        const std::complex<float>* b,
+        const std::complex<float>* W,
+        const std::complex<float>* n,
+        const float prefactor=1.0f,
         const bool gpu=true
     );
 
 #ifdef __PYTHONCC__
     Psi(
-        const xt::pytensor<std::complex<double>, 1u>& a,
-        const xt::pytensor<std::complex<double>, 1u>& b,
-        const xt::pytensor<std::complex<double>, 2u>& W,
-        const xt::pytensor<std::complex<double>, 1u>& n,
-        const double prefactor=1.0,
+        const xt::pytensor<std::complex<float>, 1u>& a,
+        const xt::pytensor<std::complex<float>, 1u>& b,
+        const xt::pytensor<std::complex<float>, 2u>& W,
+        const xt::pytensor<std::complex<float>, 1u>& n,
+        const float prefactor=1.0f,
         const bool gpu=true
     ) : gpu(gpu) {
         this->N = a.shape()[0];
@@ -311,18 +311,18 @@ public:
     }
 
     void update_params(
-        const xt::pytensor<std::complex<double>, 1u>& a,
-        const xt::pytensor<std::complex<double>, 1u>& b,
-        const xt::pytensor<std::complex<double>, 2u>& W,
-        const xt::pytensor<std::complex<double>, 1u>& n
+        const xt::pytensor<std::complex<float>, 1u>& a,
+        const xt::pytensor<std::complex<float>, 1u>& b,
+        const xt::pytensor<std::complex<float>, 2u>& W,
+        const xt::pytensor<std::complex<float>, 1u>& n
     ) {
         this->update_params(
             a.data(), b.data(), W.data(), n.data()
         );
     }
 
-    xt::pytensor<complex<double>, 1> as_vector_py() const {
-        auto result = xt::pytensor<complex<double>, 1>(
+    xt::pytensor<complex<float>, 1> as_vector_py() const {
+        auto result = xt::pytensor<complex<float>, 1>(
             std::array<long int, 1>({static_cast<long int>(pow(2, this->N))})
         );
         this->as_vector(result.data());
@@ -330,8 +330,8 @@ public:
         return result;
     }
 
-    xt::pytensor<complex<double>, 1> O_k_vector_py(const Spins& spins) const {
-        auto result = xt::pytensor<complex<double>, 1>(
+    xt::pytensor<complex<float>, 1> O_k_vector_py(const Spins& spins) const {
+        auto result = xt::pytensor<complex<float>, 1>(
             std::array<long int, 1>({static_cast<long int>(this->num_active_params)})
         );
         this->O_k_vector(result.data(), spins);
@@ -348,17 +348,17 @@ public:
     }
 
     void update_params(
-        const std::complex<double>* a,
-        const std::complex<double>* b,
-        const std::complex<double>* W,
-        const std::complex<double>* n,
+        const std::complex<float>* a,
+        const std::complex<float>* b,
+        const std::complex<float>* W,
+        const std::complex<float>* n,
         const bool ptr_on_gpu=false
     );
 
-    void as_vector(complex<double>* result) const;
-    void O_k_vector(complex<double>* result, const Spins& spins) const;
-    double norm_function(const ExactSummation& exact_summation) const;
-    std::complex<double> log_psi_s_std(const Spins& spins);
+    void as_vector(complex<float>* result) const;
+    void O_k_vector(complex<float>* result, const Spins& spins) const;
+    float norm_function(const ExactSummation& exact_summation) const;
+    std::complex<float> log_psi_s_std(const Spins& spins);
 
     unsigned int get_num_params_py() const {
         return this->get_num_params();
